refactor(orchestrator): modularize leo-protocol-orchestrator.js (1418 LOC)
Extract 8 focused modules from monolithic file:
- constants.js (70 LOC): Phase definitions and requirements
- session-decision-logger.js (52 LOC): Audit trail logging
- validation.js (258 LOC): Requirement validation and phase gates
- phase-execution.js (276 LOC): Individual phase execution methods
- prd-generation.js (308 LOC): PRD and user story generation
- compliance.js (159 LOC): Compliance reporting and retrospectives
- helpers.js (164 LOC): Initialization and utility functions
- index.js (329 LOC): Main orchestrator class

All modules under 500 LOC threshold.
Original file converted to re-export wrapper (92 LOC).
Backward compatibility maintained via comprehensive re-exports.

Part of SD-LEO-REFACTOR-LARGE-FILES-003 orchestrator.
