+++ scripts/leo-protocol-orchestrator.js
@@ -8,1412 +8,85 @@
  * This is the SINGLE ENTRY POINT for all Strategic Directive executions
  * It ensures every step is followed and nothing is missed
  *
- * v2.2.0 Changes:
- * - Integrated Evidence Pack Generator for post-session audit
- * - Evidence pack generated at session completion
- * - Records SDs touched, gates executed, decisions made
+ * REFACTORED: This file now re-exports from modular structure
+ * See leo-orchestrator/ directory for implementation:
+ *   - constants.js - Phase definitions and requirements
+ *   - session-decision-logger.js - Audit trail for decisions
+ *   - validation.js - Requirement validation and phase gates
+ *   - phase-execution.js - Individual phase execution methods
+ *   - prd-generation.js - PRD and user story generation
+ *   - compliance.js - Compliance reporting and retrospectives
+ *   - helpers.js - Initialization and utility functions
+ *   - index.js - Main orchestrator class
  *
- * v2.1.0 Changes:
- * - Integrated Session Guardian for checkpointing and safe-stop
- * - Added checkpoint after each gate validation
- * - Added safe-stop detection for dangerous commands
- * - Added loop detection for stuck operations
- *
- * v2.0.0 Changes:
- * - Removed all interactive prompts (inquirer) for extended session support
- * - Added SessionDecisionLogger for audit trail
- * - Deterministic validation with clear remediation messages
- * - Conservative defaults when ambiguous (fail safe)
+ * SD-LEO-REFACTOR-ORCH-002
  */
 
-import { createClient } from '@supabase/supabase-js';
-import dotenv from 'dotenv';
-import chalk from 'chalk';
-import fs from 'fs/promises';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import { dirname } from 'path';
-import { execSync } from 'child_process';
-import { createSessionGuardian } from './lib/session-guardian.js';
-import { createEvidencePackGenerator } from './lib/evidence-pack-generator.js';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-/**
- * SessionDecisionLogger - Records all decisions made without human input
- * Enables post-session audit of automated decisions
- */
-class SessionDecisionLogger {
-  constructor(sessionId) {
-    this.sessionId = sessionId;
-    this.decisions = [];
-    this.logPath = path.join(__dirname, '..', 'docs', 'audit', 'sessions', `${new Date().toISOString().split('T')[0]}`);
-    this.logFile = path.join(this.logPath, `session_decisions_${sessionId}.json`);
-  }
-
-  async init() {
-    await fs.mkdir(this.logPath, { recursive: true });
-  }
-
-  log(decision) {
-    const entry = {
-      timestamp: new Date().toISOString(),
-      sessionId: this.sessionId,
-      ...decision
-    };
-    this.decisions.push(entry);
-    console.log(chalk.gray(`  [DECISION] ${decision.type}: ${decision.action} - ${decision.reason}`));
-  }
-
-  async save() {
-    await fs.writeFile(this.logFile, JSON.stringify(this.decisions, null, 2));
-    return this.logFile;
-  }
-
-  getDecisions() {
-    return this.decisions;
-  }
-}
-
-dotenv.config();
-
-class LEOProtocolOrchestrator {
-  constructor() {
-    // Use service role key for full access to LEO tables (anon key blocked by RLS on some tables)
-    this.supabase = createClient(
-      process.env.NEXT_PUBLIC_SUPABASE_URL,
-      process.env.SUPABASE_SERVICE_ROLE_KEY || process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
-    );
-
-    // Define the immutable phase sequence
-    this.phases = ['LEAD', 'PLAN', 'EXEC', 'VERIFICATION', 'APPROVAL'];
-
-    // Decision logger for audit trail (initialized in executeSD)
-    this.decisionLogger = null;
-
-    // Session guardian for checkpointing and safe-stop (initialized in executeSD)
-    this.sessionGuardian = null;
-
-    // Evidence pack generator for post-session audit (initialized in executeSD)
-    this.evidencePack = null;
-
-    // Track execution state
-    this.executionState = {
-      sdId: null,
-      currentPhase: null,
-      completedPhases: [],
-      skippedSteps: [],
-      violations: [],
-      startTime: null,
-      sessionId: null
-    };
-
-    // Phase requirements (cannot be bypassed)
-    this.phaseRequirements = {
-      LEAD: [
-        'session_prologue_completed',
-        'priority_justified',
-        'strategic_objectives_defined',
-        'handoff_created_in_database',
-        'no_over_engineering_check'
-      ],
-      PLAN: [
-        'prd_created_in_database',
-        'acceptance_criteria_defined',
-        'sub_agents_activated',
-        'test_plan_created',
-        'handoff_from_lead_received'
-      ],
-      EXEC: [
-        'pre_implementation_checklist',
-        'correct_app_verified',
-        'screenshots_taken',
-        'implementation_completed',
-        'git_commit_created',
-        'github_push_completed'
-      ],
-      VERIFICATION: [
-        'all_tests_executed',
-        'acceptance_criteria_verified',
-        'sub_agent_consensus',
-        'supervisor_verification_done',
-        'confidence_score_calculated'
-      ],
-      APPROVAL: [
-        'human_approval_requested',
-        'over_engineering_rubric_run',
-        'human_decision_received',
-        'status_updated_in_database',
-        'retrospective_completed'
-      ]
-    };
-  }
-
-  /**
-   * Main execution entry point
-   * v2.2.0: Evidence Pack integration for post-session audit
-   * v2.1.0: Session Guardian integration for checkpointing and safe-stop
-   * v2.0.0: Non-interactive - no prompts, deterministic execution
-   */
-  async executeSD(sdId, options = {}) {
-    console.log(chalk.blue.bold('\n🚀 LEO PROTOCOL ORCHESTRATOR v2.2.0 (Evidence Pack)'));
-    console.log(chalk.blue('━'.repeat(50)));
-
-    try {
-      // Initialize execution
-      await this.initializeExecution(sdId);
-
-      // Initialize decision logger
-      this.decisionLogger = new SessionDecisionLogger(this.executionState.sessionId);
-      await this.decisionLogger.init();
-
-      // Initialize session guardian for checkpointing and safe-stop
-      this.sessionGuardian = await createSessionGuardian(this.executionState.sessionId);
-      this.sessionGuardian.setCurrentSD(sdId);
-
-      // Initialize evidence pack generator
-      this.evidencePack = await createEvidencePackGenerator(this.executionState.sessionId);
-      this.evidencePack.recordSD(sdId, 'started');
-
-      // Check for existing checkpoint to resume from
-      const existingCheckpoint = await this.sessionGuardian.loadCheckpoint();
-      if (existingCheckpoint && !options.force) {
-        console.log(chalk.yellow(`\n📍 Found checkpoint from ${existingCheckpoint.timestamp}`));
-        console.log(chalk.yellow(`   Last phase: ${existingCheckpoint.state?.currentPhase || 'unknown'}`));
-        console.log(chalk.yellow(`   Gates completed: ${existingCheckpoint.state?.gatesCompleted?.length || 0}`));
-        this.sessionGuardian.restoreFromCheckpoint(existingCheckpoint);
-        this.decisionLogger.log({
-          type: 'CHECKPOINT_RESTORE',
-          action: 'resumed',
-          reason: `Resuming from checkpoint at ${existingCheckpoint.timestamp}`,
-          lastPhase: existingCheckpoint.state?.currentPhase
-        });
-      }
-
-      // Step 1: Mandatory session prologue
-      await this.enforceSessionPrologue();
-
-      // Step 2: Verify SD exists and is eligible
-      await this.verifySDEligibility(sdId);
-
-      // Step 3: Execute each phase with strict gates
-      let previousPhase = null;
-      for (const phase of this.phases) {
-        console.log(chalk.yellow(`\n📋 Phase: ${phase}`));
-        console.log(chalk.gray('─'.repeat(40)));
-
-        // Check resource limits before each phase
-        const resourceCheck = this.sessionGuardian.checkResourceLimits();
-        if (resourceCheck.exceeded) {
-          throw new Error(`RESOURCE_LIMIT: ${resourceCheck.reason}`);
-        }
-
-        // Check if phase can be skipped (only if already complete)
-        const canSkip = await this.checkPhaseCompletion(sdId, phase);
-        if (canSkip && !options.force) {
-          console.log(chalk.green(`✓ ${phase} already complete`));
-          this.decisionLogger.log({
-            type: 'PHASE_SKIP',
-            phase,
-            action: 'skipped',
-            reason: 'Phase already completed'
-          });
-          continue;
-        }
-
-        // Record phase transition
-        if (previousPhase) {
-          await this.sessionGuardian.recordPhaseTransition(previousPhase, phase);
-        }
-
-        // Execute the phase
-        await this.executePhase(phase, sdId);
-
-        // Enforce phase gate (blocking)
-        await this.enforcePhaseGate(phase, sdId);
-
-        // Record gate passed and save checkpoint
-        await this.sessionGuardian.recordGatePassed(`${phase}_GATE`, {
-          sdId,
-          phase,
-          timestamp: new Date().toISOString()
-        });
-
-        // Record gate in evidence pack
-        this.evidencePack.recordGate(`${phase}_GATE`, true, { sdId, phase });
-
-        // Record phase completion
-        await this.recordPhaseCompletion(phase, sdId);
-        previousPhase = phase;
-      }
-
-      // Mark SD as completed in evidence pack
-      this.evidencePack.recordSD(sdId, 'completed');
-
-      // Step 4: Mandatory retrospective
-      await this.enforceRetrospective(sdId);
-
-      // Step 5: Final compliance report
-      await this.generateComplianceReport(sdId);
-
-      // Save decision log
-      const logFile = await this.decisionLogger.save();
-      console.log(chalk.gray(`\n📝 Decisions logged to: ${logFile}`));
-
-      // Copy decisions to evidence pack
-      for (const decision of this.decisionLogger.getDecisions()) {
-        this.evidencePack.recordDecision(decision);
-      }
-
-      // Mark session guardian as complete (clears checkpoint)
-      await this.sessionGuardian.complete();
-      console.log(chalk.gray('\n🛡️  Session guardian summary:'));
-      const summary = this.sessionGuardian.getSummary();
-      console.log(chalk.gray(`   Duration: ${summary.durationMinutes} minutes`));
-      console.log(chalk.gray(`   Operations: ${summary.totalOperations}`));
-      console.log(chalk.gray(`   Gates: ${summary.gatesCompleted}`));
-
-      // Generate evidence pack
-      const packPath = await this.evidencePack.generate();
-      console.log(chalk.gray(`\n📦 Evidence pack: ${packPath}`));
-
-      console.log(chalk.green.bold('\n✅ SD EXECUTION COMPLETE WITH 100% COMPLIANCE'));
-
-    } catch (error) {
-      console.error(chalk.red.bold('\n❌ EXECUTION FAILED:'), error.message);
-
-      // Log failure decision
-      if (this.decisionLogger) {
-        this.decisionLogger.log({
-          type: 'EXECUTION_FAILURE',
-          action: 'halted',
-          reason: error.message,
-          phase: this.executionState.currentPhase
-        });
-        await this.decisionLogger.save();
-      }
-
-      // Record failure in evidence pack and generate it
-      if (this.evidencePack) {
-        this.evidencePack.recordSD(this.executionState.sdId, 'failed');
-        this.evidencePack.recordViolation({
-          type: 'EXECUTION_FAILURE',
-          description: error.message,
-          phase: this.executionState.currentPhase
-        });
-
-        // Copy decisions to evidence pack before generating
-        if (this.decisionLogger) {
-          for (const decision of this.decisionLogger.getDecisions()) {
-            this.evidencePack.recordDecision(decision);
-          }
-        }
-
-        try {
-          const packPath = await this.evidencePack.generate();
-          console.log(chalk.gray(`\n📦 Evidence pack (failure): ${packPath}`));
-        } catch (packError) {
-          console.warn('⚠️  Could not generate evidence pack:', packError.message);
-        }
-      }
-
-      // Save guardian checkpoint for recovery
-      if (this.sessionGuardian) {
-        await this.sessionGuardian.fail(error.message);
-        console.log(chalk.yellow('\n📍 Checkpoint saved for recovery'));
-        console.log(chalk.yellow('   Restart orchestrator to resume from last checkpoint'));
-      }
-
-      await this.handleExecutionFailure(error);
-      throw error;
-    }
-  }
-
-  /**
-   * Initialize execution tracking
-   */
-  async initializeExecution(sdId) {
-    this.executionState = {
-      sdId,
-      currentPhase: null,
-      completedPhases: [],
-      skippedSteps: [],
-      violations: [],
-      startTime: new Date(),
-      sessionId: `LEO-${Date.now()}`
-    };
-
-    // Store in database for audit
-    await this.supabase.from('leo_execution_sessions').insert({
-      id: this.executionState.sessionId,
-      sd_id: sdId,
-      started_at: this.executionState.startTime,
-      status: 'in_progress'
-    });
-  }
-
-  /**
-   * Execute a command with session guardian checks
-   * Validates safe-stop patterns and tracks for loop detection
-   *
-   * @param {string} command - Command to execute
-   * @param {Object} options - execSync options
-   * @returns {Buffer|string} Command output
-   * @throws {Error} If command matches safe-stop pattern or loop detected
-   */
-  safeExec(command, options = {}) {
-    // Check safe-stop patterns
-    if (this.sessionGuardian) {
-      this.sessionGuardian.validateCommand(command);
-      this.sessionGuardian.checkLoop(`exec:${command.substring(0, 50)}`);
-    }
-
-    // Execute the command
-    return execSync(command, {
-      encoding: 'utf-8',
-      ...options
-    });
-  }
-
-  /**
-   * Track an operation for loop detection
-   * @param {string} operation - Operation identifier
-   */
-  trackOperation(operation) {
-    if (this.sessionGuardian) {
-      this.sessionGuardian.checkLoop(operation);
-    }
-  }
-
-  /**
-   * Enforce session prologue
-   */
-  async enforceSessionPrologue() {
-    console.log(chalk.cyan('\n📖 SESSION PROLOGUE CHECK'));
-
-    // Check if prologue was completed
-    const prologuePath = path.join(__dirname, '../.session-prologue-completed');
-
-    try {
-      await fs.access(prologuePath);
-      console.log(chalk.green('✓ Session prologue completed'));
-    } catch {
-      console.log(chalk.yellow('⚠️  Session prologue not found'));
-
-      // Generate and display prologue
-      console.log(chalk.gray('\n' + '='.repeat(50)));
-      console.log(chalk.white.bold('LEO PROTOCOL SESSION PROLOGUE'));
-      console.log(chalk.gray('='.repeat(50)));
-      console.log('1. Follow LEAD→PLAN→EXEC - Target ≥85% gate pass rate');
-      console.log('2. Use sub-agents - Architect, QA, Reviewer');
-      console.log('3. Database-first - No markdown files as source');
-      console.log('4. Small PRs - Keep diffs ≤100 lines');
-      console.log('5. 7-element handoffs required');
-      console.log('6. Priority-first - Use npm run prio:top3');
-      console.log(chalk.gray('='.repeat(50) + '\n'));
-
-      // Mark as completed
-      await fs.writeFile(prologuePath, new Date().toISOString());
-    }
-  }
-
-  /**
-   * Verify SD is eligible for execution
-   * v2.0.0: Non-interactive - logs decision and proceeds or fails based on rules
-   */
-  async verifySDEligibility(sdId) {
-    const { data: sd, error } = await this.supabase
-      .from('strategic_directives_v2')
-      .select('*')
-      .eq('id', sdId)
-      .single();
-
-    if (error || !sd) {
-      throw new Error(`SD ${sdId} not found. Remediation: Verify the SD ID exists in strategic_directives_v2 table.`);
-    }
-
-    // Check priority justification
-    console.log(chalk.cyan('\n🎯 PRIORITY JUSTIFICATION'));
-    console.log(`Priority: ${sd.priority || 'not set'}`);
-    console.log(`Status: ${sd.status}`);
-    console.log(`SD Type: ${sd.sd_type || 'not set'}`);
-
-    // v2.0.0: Non-interactive priority handling
-    // Low/no priority SDs are allowed but logged for audit
-    if (!sd.priority || sd.priority === 'low') {
-      this.decisionLogger.log({
-        type: 'LOW_PRIORITY_SD',
-        sdId,
-        action: 'proceed',
-        reason: 'Low/no priority SD allowed in non-interactive mode. Review in post-session audit.',
-        priority: sd.priority || 'not set'
-      });
-      console.log(chalk.yellow('⚠️  Low/no priority SD - proceeding (logged for audit)'));
-    }
-
-    // Check if SD is in valid status for execution
-    const validStatuses = ['approved', 'in_progress', 'pending', 'ready'];
-    if (!validStatuses.includes(sd.status)) {
-      throw new Error(`SD ${sdId} has status '${sd.status}' which is not valid for execution. Valid statuses: ${validStatuses.join(', ')}. Remediation: Update SD status or select a different SD.`);
-    }
-
-    // Store SD for later use
-    this.currentSD = sd;
-  }
-
-  /**
-   * Execute a specific phase
-   */
-  async executePhase(phase, sdId) {
-    this.executionState.currentPhase = phase;
-
-    switch (phase) {
-      case 'LEAD':
-        await this.executeLEADPhase(sdId);
-        break;
-      case 'PLAN':
-        await this.executePLANPhase(sdId);
-        break;
-      case 'EXEC':
-        await this.executeEXECPhase(sdId);
-        break;
-      case 'VERIFICATION':
-        await this.executeVERIFICATIONPhase(sdId);
-        break;
-      case 'APPROVAL':
-        await this.executeAPPROVALPhase(sdId);
-        break;
-    }
-  }
-
-  /**
-   * Enforce phase gate - blocking validation
-   */
-  async enforcePhaseGate(phase, sdId) {
-    console.log(chalk.yellow(`\n🚦 ${phase} PHASE GATE VALIDATION`));
-
-    const requirements = this.phaseRequirements[phase];
-    const results = {};
-    let allPassed = true;
-
-    for (const requirement of requirements) {
-      const passed = await this.validateRequirement(phase, requirement, sdId);
-      results[requirement] = passed;
-
-      if (passed) {
-        console.log(chalk.green(`  ✓ ${requirement}`));
-      } else {
-        console.log(chalk.red(`  ✗ ${requirement}`));
-        allPassed = false;
-      }
-    }
-
-    if (!allPassed) {
-      // Record violation
-      this.executionState.violations.push({
-        phase,
-        failedRequirements: Object.keys(results).filter(r => !results[r]),
-        timestamp: new Date()
-      });
-
-      // Block progression
-      throw new Error(`${phase} gate validation failed. Fix requirements and retry.`);
-    }
-
-    console.log(chalk.green.bold(`✅ ${phase} GATE PASSED`));
-  }
-
-  /**
-   * Validate a specific requirement
-   * v2.0.0: Non-interactive - deterministic validation with database checks
-   */
-  async validateRequirement(phase, requirement, sdId) {
-    // Deterministic validation based on database state
-
-    switch (requirement) {
-      case 'session_prologue_completed':
-        // Check if prologue marker file exists
-        const prologuePath = path.join(__dirname, '../.session-prologue-completed');
-        try {
-          await fs.access(prologuePath);
-          return true;
-        } catch {
-          return false;
-        }
-
-      case 'priority_justified':
-        // Check if SD has priority set or is documented as low-priority-allowed
-        return this.currentSD && (this.currentSD.priority || this.currentSD.sd_type === 'infrastructure');
-
-      case 'strategic_objectives_defined':
-        // Check if SD has objectives defined
-        return this.currentSD && (
-          this.currentSD.objectives ||
-          this.currentSD.strategic_objectives ||
-          this.currentSD.description
-        );
-
-      case 'no_over_engineering_check':
-        // Assume passed unless we have specific evidence of over-engineering
-        this.decisionLogger.log({
-          type: 'OVER_ENGINEERING_CHECK',
-          action: 'auto_pass',
-          reason: 'No automated over-engineering detection - passed by default'
-        });
-        return true;
-
-      case 'prd_created_in_database': {
-        // Only required for phases after LEAD
-        if (phase === 'LEAD') {
-          return true;
-        }
-        // Check using sd_id (correct field) not directive_id
-        const { data: prd } = await this.supabase
-          .from('product_requirements_v2')
-          .select('id')
-          .eq('sd_id', sdId)
-          .single();
-        return !!prd;
-      }
-
-      case 'acceptance_criteria_defined': {
-        const { data: prdAC } = await this.supabase
-          .from('product_requirements_v2')
-          .select('acceptance_criteria')
-          .eq('sd_id', sdId)
-          .single();
-        return prdAC && prdAC.acceptance_criteria && prdAC.acceptance_criteria.length > 0;
-      }
-
-      case 'sub_agents_activated':
-        // Check if sub-agent assignments exist for this SD
-        this.decisionLogger.log({
-          type: 'SUB_AGENT_CHECK',
-          action: 'auto_pass',
-          reason: 'Sub-agent activation handled by BMAD wrapper'
-        });
-        return true;
-
-      case 'test_plan_created': {
-        const { data: prdTest } = await this.supabase
-          .from('product_requirements_v2')
-          .select('test_scenarios')
-          .eq('sd_id', sdId)
-          .single();
-        return prdTest && prdTest.test_scenarios && prdTest.test_scenarios.length > 0;
-      }
-
-      case 'handoff_from_lead_received':
-      case 'handoff_created_in_database': {
-        const fromPhase = phase === 'PLAN' ? 'LEAD' : phase;
-        const { data: handoff } = await this.supabase
-          .from('sd_phase_handoffs')
-          .select('id')
-          .eq('sd_id', sdId)
-          .ilike('handoff_type', `%${fromPhase}%`)
-          .limit(1);
-        return handoff && handoff.length > 0;
-      }
-
-      case 'pre_implementation_checklist':
-        // Verified by EXEC phase execution
-        return true;
-
-      case 'correct_app_verified': {
-        // Check current working directory
-        try {
-          const cwd = process.cwd();
-          const isCorrect = !cwd.includes('EHG_Engineer');
-          if (!isCorrect) {
-            this.decisionLogger.log({
-              type: 'APP_VERIFICATION',
-              action: 'warning',
-              reason: `Running in EHG_Engineer directory (${cwd}) - may need to switch to EHG app`,
-              cwd
-            });
-          }
-          return true; // Don't block, just log
-        } catch {
-          return true;
-        }
-      }
-
-      case 'screenshots_taken':
-        // Cannot verify automatically - log and pass
-        this.decisionLogger.log({
-          type: 'SCREENSHOT_CHECK',
-          action: 'auto_pass',
-          reason: 'Screenshots cannot be verified automatically in non-interactive mode'
-        });
-        return true;
-
-      case 'implementation_completed':
-      case 'git_commit_created':
-      case 'github_push_completed':
-        // These are execution artifacts - assume completed if we reach gate
-        return true;
-
-      case 'all_tests_executed': {
-        // Check for recent test runs in database or CI
-        this.decisionLogger.log({
-          type: 'TEST_VERIFICATION',
-          action: 'auto_pass',
-          reason: 'Test execution verified by CI/CD pipeline'
-        });
-        return true;
-      }
-
-      case 'acceptance_criteria_verified':
-      case 'sub_agent_consensus':
-      case 'supervisor_verification_done':
-      case 'confidence_score_calculated':
-        // These are verification artifacts
-        this.decisionLogger.log({
-          type: 'VERIFICATION_CHECK',
-          requirement,
-          action: 'auto_pass',
-          reason: 'Verification handled by handoff validation system'
-        });
-        return true;
-
-      case 'human_approval_requested': {
-        const { data: approval } = await this.supabase
-          .from('leo_approval_requests')
-          .select('id')
-          .eq('sd_id', sdId)
-          .eq('status', 'pending')
-          .limit(1);
-        return approval && approval.length > 0;
-      }
-
-      case 'over_engineering_rubric_run':
-        // Assume passed - rubric runs during LEAD
-        return true;
-
-      case 'human_decision_received': {
-        const { data: decision } = await this.supabase
-          .from('leo_approval_requests')
-          .select('status')
-          .eq('sd_id', sdId)
-          .in('status', ['approved', 'rejected'])
-          .limit(1);
-        return decision && decision.length > 0;
-      }
-
-      case 'status_updated_in_database': {
-        const { data: sdStatus } = await this.supabase
-          .from('strategic_directives_v2')
-          .select('status')
-          .eq('id', sdId)
-          .single();
-        return sdStatus && sdStatus.status === 'completed';
-      }
-
-      case 'retrospective_completed': {
-        const { data: retro } = await this.supabase
-          .from('retrospectives')
-          .select('id')
-          .eq('sd_id', sdId)
-          .limit(1);
-        return retro && retro.length > 0;
-      }
-
-      default:
-        // v2.0.0: No interactive prompts - log and take conservative action
-        this.decisionLogger.log({
-          type: 'UNKNOWN_REQUIREMENT',
-          requirement,
-          action: 'auto_pass',
-          reason: `Unknown requirement '${requirement}' - passed by default in non-interactive mode. Review needed.`
-        });
-        console.log(chalk.yellow(`  ⚠️  Unknown requirement '${requirement}' - auto-passed (logged for review)`));
-        return true;
-    }
-  }
-
-  /**
-   * LEAD Phase execution
-   */
-  async executeLEADPhase(_sdId) {
-    console.log(chalk.blue('\n🎯 Executing LEAD Phase'));
-
-    // Check for over-engineering
-    console.log('Running over-engineering evaluation...');
-    // Would call lead-over-engineering-rubric.js here
-
-    // Create handoff
-    console.log('Creating LEAD→PLAN handoff...');
-    // Would create handoff in database
-  }
-
-  /**
-   * PLAN Phase execution
-   * v2.0.0: Non-interactive - PRD must exist, no auto-generation
-   */
-  async executePLANPhase(sdId) {
-    console.log(chalk.blue('\n📐 Executing PLAN Phase'));
-
-    // v2.0.0: Check if PRD already exists (using correct sd_id field)
-    const { data: existingPrd } = await this.supabase
-      .from('product_requirements_v2')
-      .select('id, title, status')
-      .eq('sd_id', sdId)
-      .single();
-
-    if (existingPrd) {
-      console.log(chalk.green(`✓ PRD exists: ${existingPrd.title} (status: ${existingPrd.status})`));
-      this.decisionLogger.log({
-        type: 'PRD_CHECK',
-        action: 'found',
-        reason: `PRD found for SD ${sdId}`,
-        prdId: existingPrd.id,
-        status: existingPrd.status
-      });
-      return;
-    }
-
-    // v2.0.0: PRD is REQUIRED - do not auto-generate placeholder PRDs
-    // PRDs must be created through the proper add-prd-to-database.js flow
-    // which ensures proper validation and quality gates
-    this.decisionLogger.log({
-      type: 'PRD_MISSING',
-      action: 'blocked',
-      reason: `No PRD found for SD ${sdId}. PRD creation blocked in non-interactive mode.`,
-      sdId
-    });
-
-    throw new Error(
-      `PLAN phase requires PRD. No PRD found for SD ${sdId}.\n` +
-      'Remediation:\n' +
-      `  1. Create PRD using: node scripts/add-prd-to-database.js ${sdId}\n` +
-      `  2. Or run PLAN→EXEC handoff: node scripts/handoff.js PLAN-TO-EXEC ${sdId}\n` +
-      '\n' +
-      'PRD auto-generation is disabled in non-interactive mode to ensure quality.'
-    );
-  }
-
-  /**
-   * EXEC Phase execution with mandatory checklist
-   * v2.0.0: Non-interactive - automated verification with logging
-   */
-  async executeEXECPhase(sdId) {
-    console.log(chalk.blue('\n💻 Executing EXEC Phase'));
-
-    // MANDATORY: Pre-implementation checklist (non-interactive)
-    console.log(chalk.yellow('\n📋 EXEC PRE-IMPLEMENTATION CHECKLIST (Automated)'));
-
-    const checklist = {
-      appVerified: false,
-      gitBranchVerified: false,
-      prdExists: false
-    };
-
-    // v2.0.0: Automated app verification
-    console.log('Verifying target application...');
-    const cwd = process.cwd();
-    checklist.appVerified = true; // Log warning but don't block
-    if (cwd.includes('EHG_Engineer')) {
-      this.decisionLogger.log({
-        type: 'APP_LOCATION',
-        action: 'warning',
-        reason: `Currently in EHG_Engineer (${cwd}). For app implementation, switch to EHG directory.`,
-        cwd
-      });
-      console.log(chalk.yellow('  ⚠️  In EHG_Engineer - verify this is correct for this SD type'));
-    } else {
-      console.log(chalk.green(`  ✓ App location: ${cwd}`));
-    }
-
-    // v2.0.0: Verify git branch
-    console.log('Verifying git branch...');
-    try {
-      const branch = execSync('git branch --show-current', { encoding: 'utf-8' }).trim();
-      checklist.gitBranchVerified = true;
-      console.log(chalk.green(`  ✓ Git branch: ${branch}`));
-      this.decisionLogger.log({
-        type: 'GIT_BRANCH',
-        action: 'verified',
-        reason: `Working on branch: ${branch}`,
-        branch
-      });
-    } catch (_err) {
-      this.decisionLogger.log({
-        type: 'GIT_BRANCH',
-        action: 'warning',
-        reason: 'Could not verify git branch'
-      });
-      checklist.gitBranchVerified = true; // Don't block
-    }
-
-    // v2.0.0: Verify PRD exists (critical)
-    console.log('Verifying PRD exists...');
-    const { data: prd } = await this.supabase
-      .from('product_requirements_v2')
-      .select('id, title')
-      .eq('sd_id', sdId)
-      .single();
-
-    if (prd) {
-      checklist.prdExists = true;
-      console.log(chalk.green(`  ✓ PRD found: ${prd.title}`));
-    } else {
-      // v2.0.0: PRD is required - fail with clear message
-      throw new Error(`EXEC phase requires PRD. No PRD found for SD ${sdId}. Remediation: Run PLAN phase first to create PRD using 'node scripts/add-prd-to-database.js'.`);
-    }
-
-    // Log checklist completion
-    this.decisionLogger.log({
-      type: 'EXEC_CHECKLIST',
-      action: 'completed',
-      reason: 'Automated pre-implementation checklist passed',
-      checklist
-    });
-
-    console.log(chalk.green('✓ Pre-implementation checklist complete (automated)'));
-  }
-
-  /**
-   * VERIFICATION Phase execution
-   */
-  async executeVERIFICATIONPhase(_sdId) {
-    console.log(chalk.blue('\n🔍 Executing VERIFICATION Phase'));
-
-    // Run supervisor verification
-    console.log('Running PLAN supervisor verification...');
-    // Would call plan-supervisor-verification.js
-  }
-
-  /**
-   * APPROVAL Phase execution with human gate
-   * v2.0.0: Non-interactive - creates approval request, checks for existing approval
-   */
-  async executeAPPROVALPhase(sdId) {
-    console.log(chalk.blue('\n✅ Executing APPROVAL Phase'));
-
-    // Check if approval already exists
-    const { data: existingApproval } = await this.supabase
-      .from('leo_approval_requests')
-      .select('id, status, approved_at')
-      .eq('sd_id', sdId)
-      .in('status', ['approved', 'pending'])
-      .order('created_at', { ascending: false })
-      .limit(1);
-
-    if (existingApproval && existingApproval.length > 0) {
-      const approval = existingApproval[0];
-      if (approval.status === 'approved') {
-        console.log(chalk.green(`✓ Approval already granted at ${approval.approved_at}`));
-        this.decisionLogger.log({
-          type: 'APPROVAL_CHECK',
-          action: 'already_approved',
-          reason: `SD already approved at ${approval.approved_at}`,
-          approvalId: approval.id
-        });
-        return;
-      } else {
-        // Pending approval - in non-interactive mode, we wait for handoff system
-        console.log(chalk.yellow('⚠️  Approval pending - will be handled by LEAD-FINAL-APPROVAL handoff'));
-        this.decisionLogger.log({
-          type: 'APPROVAL_CHECK',
-          action: 'pending',
-          reason: 'Approval request exists but pending. Use handoff.js LEAD-FINAL-APPROVAL to complete.',
-          approvalId: approval.id
-        });
-        return;
-      }
-    }
-
-    // v2.0.0: Create approval request but don't wait interactively
-    console.log(chalk.cyan('\n🛡️ Creating approval request...'));
-
-    const approvalRequest = {
-      id: `APPROVAL-${Date.now()}`,
-      sd_id: sdId,
-      requested_at: new Date(),
-      status: 'pending',
-      type: 'LEAD_FINAL_APPROVAL'
-    };
-
-    await this.supabase
-      .from('leo_approval_requests')
-      .insert(approvalRequest);
-
-    this.decisionLogger.log({
-      type: 'APPROVAL_REQUEST',
-      action: 'created',
-      reason: 'Approval request created. Complete via LEAD-FINAL-APPROVAL handoff.',
-      approvalId: approvalRequest.id
-    });
-
-    console.log(chalk.yellow(`\n📋 APPROVAL REQUEST CREATED: ${approvalRequest.id}`));
-    console.log(chalk.yellow('   Complete approval using: node scripts/handoff.js LEAD-FINAL-APPROVAL <SD-ID>'));
-    console.log(chalk.yellow('   Or approve via database: UPDATE leo_approval_requests SET status=\'approved\' WHERE id=\'...\';'));
-  }
-
-  /**
-   * Enforce retrospective
-   * v2.0.0: Non-interactive - checks for existing retrospective or creates placeholder
-   */
-  async enforceRetrospective(sdId) {
-    console.log(chalk.cyan('\n📝 RETROSPECTIVE CHECK'));
-
-    // Check if retrospective already exists
-    const { data: existingRetro } = await this.supabase
-      .from('retrospectives')
-      .select('id, created_at')
-      .eq('sd_id', sdId)
-      .limit(1);
-
-    if (existingRetro && existingRetro.length > 0) {
-      console.log(chalk.green(`✓ Retrospective already exists (created: ${existingRetro[0].created_at})`));
-      this.decisionLogger.log({
-        type: 'RETROSPECTIVE_CHECK',
-        action: 'already_exists',
-        reason: `Retrospective found for SD ${sdId}`,
-        retroId: existingRetro[0].id
-      });
-      return;
-    }
-
-    // v2.0.0: In non-interactive mode, create a placeholder retrospective
-    // or require it to be created via the proper retrospective flow
-    console.log(chalk.yellow('⚠️  No retrospective found'));
-    console.log(chalk.yellow('   Create retrospective using: node scripts/create-retrospective.js <SD-ID>'));
-
-    this.decisionLogger.log({
-      type: 'RETROSPECTIVE_REQUIRED',
-      action: 'not_created',
-      reason: 'Retrospective should be created via create-retrospective.js for proper quality validation',
-      sdId
-    });
-
-    // Don't block orchestrator completion - retrospective can be added after
-    console.log(chalk.gray('   (Retrospective creation is recommended but not blocking in non-interactive mode)'));
-  }
-
-  /**
-   * Generate compliance report
-   */
-  async generateComplianceReport(sdId) {
-    console.log(chalk.cyan('\n📊 COMPLIANCE REPORT'));
-    console.log(chalk.gray('─'.repeat(40)));
-
-    const report = {
-      sd_id: sdId,
-      session_id: this.executionState.sessionId,
-      phases_completed: this.executionState.completedPhases.length,
-      violations: this.executionState.violations.length,
-      skipped_steps: this.executionState.skippedSteps.length,
-      duration: new Date() - this.executionState.startTime,
-      compliance_score: this.executionState.violations.length === 0 ? 100 : 0
-    };
-
-    console.log(`SD: ${report.sd_id}`);
-    console.log(`Phases Completed: ${report.phases_completed}/5`);
-    console.log(`Violations: ${report.violations}`);
-    console.log(`Compliance Score: ${report.compliance_score}%`);
-    console.log(`Duration: ${Math.round(report.duration / 1000)}s`);
-
-    // Store report
-    await this.supabase
-      .from('leo_compliance_reports')
-      .insert(report);
-  }
-
-  /**
-   * Check if phase is already complete
-   */
-  async checkPhaseCompletion(sdId, phase) {
-    const { data } = await this.supabase
-      .from('leo_phase_completions')
-      .select('completed_at')
-      .eq('sd_id', sdId)
-      .eq('phase', phase)
-      .single();
-
-    return !!data?.completed_at;
-  }
-
-  /**
-   * Record phase completion
-   */
-  async recordPhaseCompletion(phase, sdId) {
-    this.executionState.completedPhases.push(phase);
-
-    await this.supabase
-      .from('leo_phase_completions')
-      .upsert({
-        sd_id: sdId,
-        phase,
-        completed_at: new Date(),
-        session_id: this.executionState.sessionId
-      });
-  }
-
-  /**
-   * Handle execution failure
-   */
-  async handleExecutionFailure(error) {
-    // Update session status
-    await this.supabase
-      .from('leo_execution_sessions')
-      .update({
-        status: 'failed',
-        failed_at: new Date(),
-        error_message: error.message,
-        failed_phase: this.executionState.currentPhase
-      })
-      .eq('id', this.executionState.sessionId);
-
-    // Log violation
-    await this.supabase
-      .from('leo_violations')
-      .insert({
-        session_id: this.executionState.sessionId,
-        sd_id: this.executionState.sdId,
-        phase: this.executionState.currentPhase,
-        violation_type: 'execution_failure',
-        details: error.message,
-        timestamp: new Date()
-      });
-  }
-
-  /**
-   * Check if SD is consolidated (has backlog items)
-   */
-  async isConsolidatedSD(sdId) {
-    const { data: items } = await this.supabase
-      .from('sd_backlog_map')
-      .select('backlog_id')
-      .eq('sd_id', sdId)
-      .limit(1);
-
-    return items && items.length > 0;
-  }
-
-  /**
-   * Fetch backlog items for consolidated SD
-   */
-  async fetchBacklogItems(sdId) {
-    const { data: items, error } = await this.supabase
-      .from('sd_backlog_map')
-      .select('*')
-      .eq('sd_id', sdId)
-      .order('stage_number', { ascending: true });
-
-    if (error) {
-      console.log(chalk.yellow(`⚠️  No backlog items found for ${sdId}`));
-      return [];
-    }
-
-    return items || [];
-  }
-
-  /**
-   * Generate PRD from Strategic Directive
-   */
-  async generatePRD(sd) {
-    const prdId = `PRD-${sd.id}-${Date.now()}`;
-
-    // Check if this is a consolidated SD
-    const isConsolidated = await this.isConsolidatedSD(sd.id);
-    let backlogItems = [];
-
-    if (isConsolidated) {
-      console.log(chalk.cyan('   Detected consolidated SD - fetching backlog items...'));
-      backlogItems = await this.fetchBacklogItems(sd.id);
-      console.log(chalk.green(`   Found ${backlogItems.length} backlog items`));
-    }
-
-    // Generate PRD content structure
-    const prdContent = {
-      version: '1.0.0',
-      product_overview: {
-        name: sd.title,
-        description: sd.description || 'Implementation of strategic directive objectives',
-        target_users: ['Internal stakeholders', 'End users'],
-        business_value: sd.business_value || 'Achieve strategic objectives'
-      },
-      user_stories: isConsolidated ?
-        this.generateUserStoriesFromBacklog(backlogItems, sd) :
-        this.generateUserStories(sd),
-      technical_requirements: {
-        architecture: 'To be defined based on requirements',
-        technology_stack: 'Existing stack',
-        integrations: [],
-        performance_requirements: {
-          response_time: '< 2s',
-          availability: '99.9%'
-        }
-      },
-      acceptance_criteria: this.generateAcceptanceCriteria(sd),
-      test_plan: {
-        unit_tests: 'Required for all new functionality',
-        integration_tests: 'Required for API endpoints',
-        user_acceptance_tests: 'Required before deployment',
-        performance_tests: 'As needed based on requirements'
-      },
-      metadata: {
-        generated_by: 'LEO Protocol Orchestrator',
-        generation_method: 'automated_from_sd',
-        sd_priority: sd.priority || 'medium',
-        is_consolidated: isConsolidated,
-        backlog_item_count: backlogItems.length
-      }
-    };
-
-    // Add backlog evidence appendix if consolidated
-    if (isConsolidated) {
-      prdContent.backlog_evidence = this.generateBacklogEvidence(backlogItems);
-
-      // Validate all items are included
-      const expectedCount = sd.metadata?.item_count || sd.total_items || 0;
-      if (expectedCount > 0 && backlogItems.length !== expectedCount) {
-        console.log(chalk.yellow(`⚠️  Item count mismatch: Found ${backlogItems.length}, expected ${expectedCount}`));
-      }
-    }
-
-    // Return PRD record with content as JSON string
-    return {
-      id: prdId,
-      sd_id: sd.id,
-      title: `PRD: ${sd.title}`,
-      content: JSON.stringify(prdContent, null, 2),
-      status: 'draft',
-      created_at: new Date().toISOString(),
-      updated_at: new Date().toISOString(),
-      metadata: {
-        generated_by: 'LEO Protocol Orchestrator',
-        sd_priority: sd.priority || 'medium',
-        is_consolidated: isConsolidated,
-        backlog_items: isConsolidated ? backlogItems.map(i => i.backlog_id) : []
-      }
-    };
-  }
-
-  /**
-   * Generate backlog evidence appendix
-   */
-  generateBacklogEvidence(backlogItems) {
-    const evidence = {};
-
-    backlogItems.forEach(item => {
-      evidence[item.backlog_id] = {
-        title: item.backlog_title,
-        priority: item.priority,
-        stage: item.stage_number,
-        category: item.extras?.Category || 'General',
-        description: item.extras?.Description_1 || item.item_description || '',
-        new_module: item.new_module || false,
-        completion_status: item.completion_status || 'NOT_STARTED',
-        raw_data: {
-          description_raw: item.description_raw,
-          my_comments: item.my_comments,
-          extras: item.extras || {}
-        }
-      };
-    });
-
-    return evidence;
-  }
-
-  /**
-   * Generate user stories from backlog items (for consolidated SDs)
-   */
-  generateUserStoriesFromBacklog(backlogItems, sd) {
-    const stories = [];
-
-    // Map backlog priority to user story priority
-    const priorityMap = {
-      'Very High': 'CRITICAL',
-      'High': 'HIGH',
-      'Medium': 'MEDIUM',
-      'Low': 'LOW',
-      'Very Low': 'LOW'
-    };
-
-    backlogItems.forEach((item, index) => {
-      // Extract description from extras if available
-      const description = item.extras?.Description_1 || item.item_description || item.description_raw || '';
-      const category = item.extras?.Category || 'General';
-
-      stories.push({
-        id: `US-${sd.id}-${String(index + 1).padStart(3, '0')}`,
-        title: item.backlog_title,
-        description: description,
-        acceptance_criteria: this.generateAcceptanceCriteriaFromBacklogItem(item),
-        priority: priorityMap[item.priority] || 'MEDIUM',
-        metadata: {
-          backlog_id: item.backlog_id,
-          stage: item.stage_number,
-          category: category,
-          new_module: item.new_module || false,
-          completion_status: item.completion_status || 'NOT_STARTED'
-        }
-      });
-    });
-
-    // Validate we got all items
-    if (stories.length !== backlogItems.length) {
-      console.log(chalk.yellow(`⚠️  Story count mismatch: ${stories.length} stories from ${backlogItems.length} items`));
-    }
-
-    return stories;
-  }
-
-  /**
-   * Generate acceptance criteria from a backlog item
-   */
-  generateAcceptanceCriteriaFromBacklogItem(item) {
-    const criteria = [];
-
-    // Basic implementation criteria
-    criteria.push(`${item.backlog_title} is fully implemented`);
-
-    // Add specific criteria based on item details
-    if (item.new_module) {
-      criteria.push('New module is created and integrated');
-    }
-
-    if (item.extras?.Description_1) {
-      // Extract key requirements from description
-      if (item.extras.Description_1.includes('integration') || item.extras.Description_1.includes('integrate')) {
-        criteria.push('Integration is tested and working');
-      }
-      if (item.extras.Description_1.includes('API')) {
-        criteria.push('API endpoints are documented and tested');
-      }
-      if (item.extras.Description_1.includes('sync')) {
-        criteria.push('Synchronization is verified bi-directionally');
-      }
-    }
-
-    // Standard criteria
-    criteria.push('Feature passes all tests');
-    criteria.push('Documentation is updated');
-
-    if (item.priority === 'Very High' || item.priority === 'High') {
-      criteria.push('Performance benchmarks are met');
-      criteria.push('Security review completed');
-    }
-
-    return criteria;
-  }
-
-  /**
-   * Generate user stories from SD objectives
-   */
-  generateUserStories(sd) {
-    const stories = [];
-    const objectives = sd.objectives || [];
-
-    // Parse objectives if they're a string
-    const objectiveList = typeof objectives === 'string'
-      ? objectives.split('\n').filter(o => o.trim())
-      : objectives;
-
-    objectiveList.forEach((objective, index) => {
-      stories.push({
-        id: `US-${sd.id}-${index + 1}`,
-        title: objective.trim(),
-        description: `As a user, I want ${objective.trim().toLowerCase()}`,
-        acceptance_criteria: [
-          `${objective.trim()} is implemented`,
-          'Feature is tested and verified',
-          'Documentation is updated'
-        ],
-        priority: 'HIGH'
-      });
-    });
-
-    // Add default story if no objectives
-    if (stories.length === 0) {
-      stories.push({
-        id: `US-${sd.id}-001`,
-        title: 'Implement strategic directive',
-        description: `As a stakeholder, I want ${sd.title} to be implemented`,
-        acceptance_criteria: [
-          'All requirements are met',
-          'Solution is tested and verified',
-          'Documentation is complete'
-        ],
-        priority: 'HIGH'
-      });
-    }
-
-    return stories;
-  }
-
-  /**
-   * Generate acceptance criteria from SD
-   */
-  generateAcceptanceCriteria(_sd) {
-    return [
-      'All user stories are implemented and tested',
-      'Code passes all quality checks (linting, type checking)',
-      'Unit test coverage meets minimum requirements',
-      'Integration tests pass successfully',
-      'Documentation is updated',
-      'Performance requirements are met',
-      'Security requirements are satisfied',
-      'User acceptance criteria are validated'
-    ];
-  }
-}
-
-// CLI execution
+// Re-export everything from modular structure
+export { default, LEOProtocolOrchestrator } from './leo-orchestrator/index.js';
+
+// Re-export constants for direct access
+export {
+  PHASES,
+  PHASE_REQUIREMENTS,
+  VALID_STATUSES,
+  PRIORITY_MAP
+} from './leo-orchestrator/constants.js';
+
+// Re-export session decision logger
+export { SessionDecisionLogger } from './leo-orchestrator/session-decision-logger.js';
+
+// Re-export validation functions
+export {
+  validateRequirement,
+  enforcePhaseGate
+} from './leo-orchestrator/validation.js';
+
+// Re-export phase execution functions
+export {
+  executePhase,
+  executeLEADPhase,
+  executePLANPhase,
+  executeEXECPhase,
+  executeVERIFICATIONPhase,
+  executeAPPROVALPhase
+} from './leo-orchestrator/phase-execution.js';
+
+// Re-export PRD generation functions
+export {
+  generatePRD,
+  isConsolidatedSD,
+  fetchBacklogItems,
+  generateBacklogEvidence,
+  generateUserStoriesFromBacklog,
+  generateUserStories,
+  generateAcceptanceCriteria,
+  generateAcceptanceCriteriaFromBacklogItem
+} from './leo-orchestrator/prd-generation.js';
+
+// Re-export compliance functions
+export {
+  checkPhaseCompletion,
+  recordPhaseCompletion,
+  generateComplianceReport,
+  handleExecutionFailure,
+  enforceRetrospective
+} from './leo-orchestrator/compliance.js';
+
+// Re-export helper functions
+export {
+  initializeExecution,
+  safeExec,
+  trackOperation,
+  enforceSessionPrologue,
+  verifySDEligibility
+} from './leo-orchestrator/helpers.js';
+
+// CLI execution - delegate to modular index
 if (import.meta.url === `file://${process.argv[1]}`) {
-  const args = process.argv.slice(2);
-
-  // Handle --help flag
-  if (args.includes('--help') || args.includes('-h')) {
-    console.log(`
-LEO Protocol Orchestrator v2.2.0
-
-Usage: node leo-protocol-orchestrator.js <SD-ID> [options]
-
-Arguments:
-  SD-ID           Strategic Directive ID to execute (e.g., SD-UAT-001)
-
-Options:
-  --force         Force re-execution of completed phases
-  --help, -h      Show this help message
-
-Features (v2.2.0):
-  - Non-interactive mode (no prompts)
-  - Session Guardian (checkpoint, safe-stop, loop detection)
-  - Evidence Pack generation for post-session audit
-  - Decision audit trail
-
-Examples:
-  node leo-protocol-orchestrator.js SD-UAT-001
-  node leo-protocol-orchestrator.js SD-UAT-001 --force
-
-Related Commands:
-  npm run sd:next        Show SD queue and recommendations
-  npm run sd:status      Show progress vs baseline
-  node scripts/handoff.js LEAD-TO-PLAN <SD-ID>
-`);
-    process.exit(0);
-  }
-
-  const orchestrator = new LEOProtocolOrchestrator();
-
-  const sdId = args[0];
-  const options = {
-    force: args.includes('--force')
-  };
-
-  if (!sdId) {
-    console.error(chalk.red('Usage: node leo-protocol-orchestrator.js <SD-ID>'));
-    console.error(chalk.gray('       Run with --help for more information'));
-    process.exit(1);
-  }
-
-  orchestrator.executeSD(sdId, options)
-    .then(() => process.exit(0))
-    .catch(() => process.exit(1));
+  import('./leo-orchestrator/index.js').then(_module => {
+    // CLI handled by index.js
+  }).catch(_err => {
+    // Error handled by index.js
+  });
 }
-
-export default LEOProtocolOrchestrator;
